requirements.txt
django==2.2.10
Django-crispy-forms
django-filter
django-widget-tweaks